tests: use "log" variable
Many tests assigned the Python logger to a variable named "logger", and
then assigned that to another variable "log". The test suites used "log"
and ignored "logger".

tests/mgr/test_ceph_83573397.py was unique in that it used the "logger"
variable instead of "log".

Standardize on the "log" variable everywhere. Directly assign the logger
to "log" and remove the unused "logger" variable.

The purpose of this change is to simplify the code and make it easier to
understand.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>