WIP: cephadm: better error when running without cdn_credentials
cephadm requires CDN credentials. If a user did not set cdn_credentials
in ~/.cephci.yaml, bootstrap() would fail with an error:

    user=cdn_cred.get("username"),
    AttributeError: 'NoneType' object has no attribute 'get'

TODO: reduce the scope of this commit so that it only raises KeyError
when CDN credentials are not found.

Rewrite the code to use a list when constructing cmd so that it is
easier to dynamically choose what cephadm bootstrap arguments we will
invoke.

TODO: add unit tests here.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>