WIP: utils: handle fully-commented ~/.cephci.yaml
If a user has not un-commented any key/values in ~/.cephci.yaml,
yaml.safe_load() returns None. This leads to TypeErrors later on when
the rest of cephci treats this value as a dictionary and tries to load
values from it with .get(), etc.

If PyYAML returns None, it means the user has an empty configuration
file. Return an empty dict to satisfy all the callers' expectations.

TODO: add unit tests here.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>